+++ packages/mongodb-chatbot-server/src/routes/conversations/createConversation.ts
@@ -7,7 +7,7 @@ import {
   ConversationCustomData,
   ConversationsService,
   SystemMessage,
-} from "../../services/ConversationsService";
+} from "mongodb-rag-core/ConversationsService";
 import {
   ApiConversation,
   convertConversationFromDbToApi,
