+++ packages/mongodb-chatbot-server/src/routes/conversations/rateMessage.test.ts
@@ -6,7 +6,7 @@ import {
   Message,
   AssistantMessage,
   ConversationsService,
-} from "../../services/ConversationsService";
+} from "mongodb-rag-core/ConversationsService";
 import { Express } from "express";
 import { ObjectId } from "mongodb-rag-core";
 import { makeRateMessageRoute } from "./rateMessage";
