+++ packages/mongodb-chatbot-server/src/routes/conversations/utils.ts
@@ -1,6 +1,6 @@
 import { isIP } from "net";
 import { Address6 } from "ip-address";
-import { Conversation, Message } from "../../services/ConversationsService";
+import { Conversation, Message } from "mongodb-rag-core/ConversationsService";
 import { References } from "mongodb-rag-core";
 import { z } from "zod";
