+++ packages/mongodb-chatbot-server/src/test/testPreProcessor/makePreprocessMongoDbUserQuery.ts
@@ -10,7 +10,7 @@ import {
   makeTypeChatJsonTranslateFunc,
   AzureOpenAiServiceConfig,
 } from "mongodb-rag-core";
-import { Message } from "../../services/ConversationsService";
+import { Message } from "mongodb-rag-core/ConversationsService";
 
 /**
   Query preprocessor that uses the Azure OpenAI service to preprocess
