+++ packages/mongodb-rag-core/package.json
@@ -48,7 +48,7 @@
     "eslint-config-prettier": "^8",
     "eslint-plugin-prettier": "^4",
     "eslint-plugin-tsdoc": "^0.2.14",
-    "express": "^4.18.2",
+    "express": "^4.19.2",
     "jest": "^29",
     "mongodb-memory-server": "^8",
     "prettier": "^2",
@@ -60,8 +60,8 @@
   },
   "dependencies": {
     "@azure/openai": "^1.0.0-beta.5",
-    "@langchain/openai": "^0.2.1",
     "@langchain/core": "^0.2.9",
+    "@langchain/openai": "^0.2.1",
     "common-tags": "^1",
     "dotenv": "^16.3.1",
     "exponential-backoff": "^3.1.1",
