+++ packages/mongodb-rag-core/src/LangchainChatLlm.ts
