+++ packages/mongodb-rag-core/src/OpenAiEmbedder.ts
@@ -1,6 +1,6 @@
 import { OpenAIClient } from "@azure/openai";
 import { Embedder } from "./Embedder";
-import { logger } from "./services/logger";
+import { logger } from "./logger";
 import { stripIndent } from "common-tags";
 import { backOff, BackoffOptions } from "exponential-backoff";
