+++ packages/mongodb-rag-core/src/index.ts
@@ -1,19 +1,27 @@
 export * from "./BoostOnAtlasSearchFilter";
+export * from "./ChatLlm";
+export * from "./ConversationsService";
 export * from "./CoreEnvVars";
 export * from "./DatabaseConnection";
+export * from "./DataStreamer";
 export * from "./DefaultFindContent";
 export * from "./DefaultFindVerifiedAnswer";
 export * from "./EmbeddedContent";
 export * from "./Embedder";
 export * from "./FindContentFunc";
 export * from "./FindVerifiedAnswerFunc";
+export * from "./LangchainChatLlm";
 export * from "./LangChainEmbedder";
+export * from "./logger";
+export * from "./MongoDbConversations";
 export * from "./MongoDbEmbeddedContentStore";
 export * from "./MongoDbVerifiedAnswerStore";
 export * from "./MongoDbPageStore";
+export * from "./OpenAiChatLlm";
 export * from "./OpenAiEmbedder";
 export * from "./Page";
 export * from "./PageFormat";
+export * from "./References";
 export * from "./TypeChatJsonTranslateFunc";
 export * from "./VectorStore";
 export * from "./VerifiedAnswer";
@@ -22,8 +30,6 @@ export * from "./arrayFilters";
 export * from "./assertEnvVars";
 export * from "./extractFrontMatter";
 export * from "./removeFrontMatter";
-export * from "./services/conversations";
-export * from "./services/logger";
 export * from "./updateFrontMatter";
 
 // Everyone share the same mongodb driver version
